Normalize bone weights
Extended the following methods by an optional `bool aNormalizeBoneWeights`
parameter (default is false):
- `get_bone_weights`
- `create_bone_weights_buffer`
- `model_t::bone_weights_for_mesh`
- `model_t::bone_weights_for_meshes`

When set to true, the returned bone weights are guaranteed to add up to
one, even if the original weights summed to > 1 (Blender does that).
If there are more than 4 bones affecting a vertex, all the weights
except the 3 highest ones are assigned to the fourth bone.