aiProcess_JoinIdenticalVertices sometimes fails silently
Allow a configuration for the assimp importer to discard the baked normals of a mesh and recalculate the normals with assimp instead. aiProcess_JoinIdenticalVertices is not applied if there are normals already defined in the model file. This removes 2/3 of the vertices of many loaded models.

See https://github.com/assimp/assimp/issues/407#issuecomment-61445211